Extract turn_helpers.rs from turn.rs (parent #680)
Move the module-level free functions and shared types out of `turn.rs`
into a focused sibling module so `turn.rs` keeps shrinking toward
responsibility-focused units (parent meta-issue #680).

Four free fns + one struct + one type alias extracted:

- `extract_filename_with_suffix` (token extractor for tool-call
  argument / feedback recovery heuristics)
- `write_stdout_rendered` (`raw_mode_safe_text` + lock + write +
  trailing-newline helper)
- `tool_result_failed` (Error-prefix or interrupted=true marker)
- `quality_confirm_cache_key` (Issue #580 SSoT memoization hash)
- `RetrievalInjection` (Issue #555 photon retrieval message +
  selected-id pair)
- `WrittenScaffoldArtifacts` (deterministic scaffold installer return
  type alias)

Behavior unchanged: byte-for-byte identical function bodies, same
field shapes for the struct and alias.

`pub(super)` limited / no facade re-export (DR3-001). Callers updated
in `actor_loop_flow.rs`, `agent_misc.rs`, `anti_pattern_flow.rs`,
`build_request_messages.rs`, `case_record_flow.rs`,
`classify_confirm_flow.rs`, `repair_job_dispatch.rs`,
`scaffold_pipeline.rs`, `streaming_reply.rs`, `truncate_tests.rs`,
and `verifier_orchestration.rs` (~30 import / call sites).

LOC delta:
- turn.rs: 175 → 99 (-76)
- new module: +103

3,114 lib tests pass; `cargo fmt` / `cargo clippy --lib --all-targets
-- -D warnings` clean.
